requirements.txt
flask
geocoder
folium
flask_restful
pyunsplash
Pillow
transformers
torch